added a README and a .gitignore.
My convention is to put my build products in a subdir named `objdir`
or `objdir-SUFFIX` for some SUFFIX, so that's the most important thing
being ignored here.  (Note however that I think currently the
`Makefile`, as a symlink, is actually part of the git repo which
unfortunately means that `objdir` is also part of the git repo,
whoops.