Further hacking, added code attempting to init the Rust runtime.
Realized of course I need to init the Rust runtime in order to use
rustc, which depends on details like task-local storage.  Not sure how
best to actually initialize the runtime from Racket.

It is also not clear if we can actually fail cleanly or even re-enter
Racket itself.  Currently a fail (which is what one expects from the
extern rustc invoking smoke test since I have not figured out how to
point it at the stdlib) causes the racket process to crash.